Split table to display active and inactive reasons
1. Active and inactive reasons can be sorted separately.
2. Inactive reasons is below active reason.
3. Assume the 'n' part of  'n filters have caught m posts' should still
display total number of reasons instead of number of active reasons.

Note: the column widths of the two tables are not guaranteed to be the
same since the widths depend on the length of contents for active and
inactive reasons, respectively.  From a practical standpoint, the width
should normalize to be about the same given enough content.