Dockerfile: fix build failures after merge
Merged master and found that docker build would no longer run.

It was complaining that my bundler version was too old, so as
suggested in the error message, we now separately install an
upgrade for bundler before attempting to run it.

Gem installation was failing with an ugly UTF-8 error.
What ultimately worked was to include a RUBYOPT environment variable
in the Dockerfile.

Credit: https://stackoverflow.com/a/32915504/874188