Attempted to move JS libs to webpack
Moves jquery, bootstrap, moment to their respective npm modules. This is an attempt to fix the
issue wherein we're loading jQuery twice, _somehow_, and causing (a) duplicate requests all
over the place, and (b) Bootstrap methods like $().modal() to error out because they're using
the wrong copy of jQuery.

This isn't fixed. This is an attempt to move, but we still get issues - I haven't tested prod,
but in dev we get "jQuery is not defined" o.e. from bootstrap - I asked a question about this on
SO, see https://stackoverflow.com/q/52467689/3160466.