Always get the revision_count
We provide the revision_count via the MS API, so should always try to fetch it. Doing so is
already a dependency to autoflagging, so doing it earlier doesn't change if we autoflag.
Fetching it earlier saves us the work of processing the flagging conditions for the cases
where the number of revisions disqualifies the post.