Don't limit check run and suite reporting to master and PRs.
It turns out that even though these two WebHooks include an array
of objects to describe any PRs which the check is associated with,
the array only contains data when the check is being performed upon
PR creation. It is not present when additional commits are added
to a PR.

There does not appear to be a way to differentiate that the commit
is currently involved in a PR based on the information provided
in these WebHooks. So, we can either have something more complex,
that includes fetching and/or storing more data, or we can either
A) report all check run failures and all check suite successes,
or we can B) limit it to only specific branches. It's already
limited to some branches based on what the GitHub Action config
is.

So, overall, it looks to me like it's better to do (A) report all
check run failures and all check suite successes which we receive.