Prevent quick duplicate/similar AJAX requests.
There are several situations, both programmatic and via user input, where
we're experiencing the effects of having duplicate requests being sent causing
unexpected operation. The one which was clearest to me was the add new domain link
dialog. While this change also resolved preventing that duplication, that specific
duplication was resolved specific to using data-remote in a <form>.