Add regex_cache and end post_matches? early
This should improve performance for matching posts
against a spam wave. In addition to immediately
returning false upon failure to match the non-regex
based criteria, a regex_cache is used to store the
compiled Regexp for each regex. This should prevent
the need to compile each of the three spam wave's
regexes each time we're testing a post.

This should substantially reduce the time it
takes to validate a spam wave and should reduce
how long it takes to process the creation
of every report.