Account for malformed access_token values
2023-08-10 SE started sending
  <token>&scope=write_access%20no_expiry
for the access_token, at least for the implicit OAuth 2.0 flow instead of the expected
  <token>

The change is causing it to be impossible to log into MS.

This isn't yet verified for the explicit OAuth 2.0 flow which MS is using. Making this change
should, effectively, verify that it's happening the same way with the explicit flow.