TYPO: Exploder -> Explorer
I think this is a typo. I'm not sure if it's a typo or intentional, but I've always found it
confusing.
It looks like it was introduced back on 2019-03-18 in commit
c98b7585fbf535852e886d598be4ca0a45a848ae, and then just continued/copied under the assumption
that it was intentional.