Remove use of DOMNodeInserted
Chrome removed DOMNodeInserted. We currently explicitly use that event
only to show the appropriate tab for post bodies. We already have code
that executes when post bodies are added to the DOM in order to render
the post preview. The code to detect when a post body is added to the
DOM is effectively duplicated here to select the correct tab.

NOTE: This is not as generically effective as a MutationObserver watching
for all DOM node insertions in the <body> would be, but is substantially
more efficient. The main drawback is that if a new way to add a post body
is added to the code base, it may be necessary to add additional way(s) to
detect that the post body has been added to the DOM (such as the special
case used for the review pages).